refactor: restructure commands — /stats for analytics, /raid for archiving
Move stats/status/attendance/suggest subcommands from /raid to /stats
and register the /stats command. Move archive/unarchive/search from
dead stats.ts into /raid. Remove notes command entirely. Clean up
unused analytics imports from raid.ts. Fix stale /stats archive
reference in config.ts. Update feature manifest to reflect new
command structure with accurate subcommand counts and docs URLs.
